backend/pyproject.toml
[project]
name = "postnatal-pulse"
version = "0.1.0"
description = "Postnatal Pulse backend"
authors = [
    { name = "jjovalle99" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi[standard]>=0.136.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "psycopg[binary,pool]>=3.3.3",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "sse-starlette>=3.3.4",
    "ulid-py>=1.1.0",
    "weasyprint>=68.1",
    "twilio>=9.10.5",
    "speechmatics-rt>=1.0.0",
    "thymia-sentinel>=1.1.0",
]

[project.scripts]
postnatal-pulse = "postnatal_pulse.main:main"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx-ws>=0.9.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
    "ty>=0.0.31",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "ANN"]
ignore = ["ANN401"]

[tool.fastapi]
entrypoint = "postnatal_pulse.main:app"
